requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
yfinance>=0.1.70
alpha-vantage>=2.3.1
fredapi>=0.4.3
quandl>=3.6.1
requests>=2.26.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
ta>=0.7.0
arch>=5.0.0
hmmlearn>=0.2.7
statsmodels>=0.13.0
pytest>=6.2.0
jupyter>=1.0.0
tqdm>=4.62.0
python-dotenv>=0.19.0
pyyaml>=6.0